lib: move get_kani_list and check functions to kani_list module
